+++ src/backend/backend_web.rs
@@ -1,22 +1,17 @@
-use std::{
-    collections::{BTreeMap, HashMap},
-    sync::Arc,
-};
+use std::sync::Arc;
 
-use anyhow::{bail, Context};
-use js_sys::{JsString, Object, Reflect, Uint8Array, WebAssembly};
+use anyhow::Context;
+use js_sys::{JsString, Reflect, Uint8Array, WebAssembly};
 use wasm_bindgen::{JsCast, JsValue};
 
-use super::{
-    AsContextMut, Export, TableType, Value, WasmEngine, WasmExternRef, WasmInstance, WasmModule,
-};
+use super::{AsContextMut, TableType, Value, WasmEngine, WasmExternRef, WasmModule};
 use crate::{
     backend::Extern,
     web::{
         conversion::{FromJs, FromStoredJs, ToJs, ToStoredJs},
         module::{self, ParsedModule},
-        Engine, Func, Global, Import, Instance, InstanceInner, JsErrorMsg, Memory, Module,
-        ModuleInner, Store, StoreContext, StoreContextMut, StoreInner, Table,
+        Engine, Func, Global, Import, Instance, JsErrorMsg, Memory, Module, ModuleInner, Store,
+        StoreContext, StoreContextMut, StoreInner, Table,
     },
     ExternType, GlobalType, ImportType, MemoryType, ValueType,
 };
@@ -41,211 +36,6 @@ impl WasmEngine for Engine {
     type StoreContextMut<'a, T: 'a> = StoreContextMut<'a, T>;
 
     type Table = Table;
-}
-
-impl WasmInstance<Engine> for Instance {
-    fn new(
-        mut store: impl super::AsContextMut<Engine>,
-        module: &Module,
-        imports: &super::Imports<Engine>,
-    ) -> anyhow::Result<Self> {
-        let _span = tracing::debug_span!("Instance::new").entered();
-        let store: &mut StoreInner<_> = &mut *store.as_context_mut();
-
-        let instance;
-        let parsed;
-        let imports_object;
-
-        {
-            let mut engine = store.engine.borrow_mut();
-            let module = &mut engine.modules[module.id];
-            parsed = module.parsed.clone();
-
-            imports_object = create_imports_object(store, imports);
-
-            // TODO: async instantiation, possibly through a `.ready().await` call on the returned
-            // module
-            // let instance = WebAssembly::instantiate_module(&module.module, &imports);
-            instance = WebAssembly::Instance::new(&module.module, &imports_object)
-                .map_err(JsErrorMsg::from)
-                .with_context(|| "Failed to instantiate module")?;
-        };
-
-        let _span = tracing::debug_span!("get_exports").entered();
-
-        let js_exports = Reflect::get(&instance, &"exports".into()).expect("exports object");
-        let exports = process_exports(store, js_exports, &parsed)?;
-
-        let instance = InstanceInner { instance, exports };
-
-        let instance = store.insert_instance(instance);
-
-        Ok(instance)
-    }
-
-    fn exports(
-        &self,
-        store: impl super::AsContext<Engine>,
-    ) -> Box<dyn Iterator<Item = super::Export<Engine>>> {
-        // TODO: modify this trait to make the lifetime of the returned iterator depend on the
-        // anonymous lifetime of the store
-        let instance: &InstanceInner = &store.as_context().instances[self.id];
-        Box::new(
-            instance
-                .exports
-                .iter()
-                .map(|(name, value)| Export {
-                    name: name.into(),
-                    value: value.clone(),
-                })
-                .collect::<Vec<_>>()
-                .into_iter(),
-        )
-    }
-
-    fn get_export(
-        &self,
-        store: impl super::AsContext<Engine>,
-        name: &str,
-    ) -> Option<super::Extern<Engine>> {
-        let instance: &InstanceInner = &store.as_context().instances[self.id];
-        instance.exports.get(name).cloned()
-    }
-}
-
-/// Creates the js import map
-fn create_imports_object<T>(store: &StoreInner<T>, imports: &super::Imports<Engine>) -> Object {
-    let _span = tracing::debug_span!("process_imports").entered();
-
-    let imports = imports
-        .into_iter()
-        .map(|((module, name), imp)| {
-            tracing::trace!(?module, ?name, ?imp, "import");
-            let js = imp.to_stored_js(store);
-
-            tracing::trace!(module, name, "export");
-
-            (module, (JsString::from(&*name), js))
-        })
-        .fold(BTreeMap::<String, Vec<_>>::new(), |mut acc, (m, value)| {
-            acc.entry(m).or_default().push(value);
-            acc
-        });
-
-    imports
-        .iter()
-        .map(|(module, imports)| {
-            let obj = Object::new();
-
-            for (name, value) in imports {
-                Reflect::set(&obj, name, value).unwrap();
-            }
-
-            (module, obj)
-        })
-        .fold(Object::new(), |acc, (m, imports)| {
-            Reflect::set(&acc, &(m).into(), &imports).unwrap();
-            acc
-        })
-}
-
-/// Processes a wasm module's exports into a hashmap
-fn process_exports<T>(
-    store: &mut StoreInner<T>,
-    exports: JsValue,
-    parsed: &ParsedModule,
-) -> anyhow::Result<HashMap<String, Extern<Engine>>> {
-    let _span = tracing::debug_span!("process_exports", ?exports).entered();
-    if !exports.is_object() {
-        bail!(
-            "WebAssembly exports must be an object, got '{:?}' instead",
-            exports
-        );
-    }
-
-    let exports: Object = exports.into();
-    let names = Object::get_own_property_names(&exports);
-    let len = names.length();
-
-    tracing::debug!(?names, ?exports);
-
-    Object::entries(&exports)
-        .into_iter()
-        .map(|entry| {
-            let name = Reflect::get_u32(&entry, 0)
-                .unwrap()
-                .as_string()
-                .expect("name is string");
-
-            let value: JsValue = Reflect::get_u32(&entry, 1).unwrap();
-
-            let _span = tracing::trace_span!("process_export", ?name, ?value).entered();
-
-            let ty = value.js_typeof();
-
-            let signature = parsed.exports.get(&name).expect("export signature").clone();
-
-            let ext = match &value
-                .js_typeof()
-                .as_string()
-                .expect("typeof returns a string")[..]
-            {
-                "function" => {
-                    let func = Func::from_exported_function(
-                        &name,
-                        store,
-                        value,
-                        signature.try_into_func().unwrap(),
-                    )
-                    .unwrap();
-
-                    Extern::Func(func)
-                }
-                "object" => {
-                    if value.is_instance_of::<js_sys::Function>() {
-                        let func = Func::from_exported_function(
-                            &name,
-                            store,
-                            value,
-                            signature.try_into_func().unwrap(),
-                        )
-                        .unwrap();
-
-                        Extern::Func(func)
-                    } else if value.is_instance_of::<WebAssembly::Table>() {
-                        let table = Table::from_stored_js(
-                            store,
-                            value,
-                            signature.try_into_table().unwrap(),
-                        )
-                        .unwrap();
-
-                        Extern::Table(table)
-                    } else if value.is_instance_of::<WebAssembly::Memory>() {
-                        let memory = Memory::from_stored_js(store, value).unwrap();
-
-                        Extern::Memory(memory)
-                    } else if value.is_instance_of::<WebAssembly::Global>() {
-                        let global = Global::from_exported_global(
-                            store,
-                            value,
-                            signature.try_into_global().unwrap(),
-                        )
-                        .unwrap();
-
-                        Extern::Global(global)
-                    } else {
-                        tracing::error!("Unsupported export type {value:?}");
-                        panic!("Unsupported export type {value:?}")
-                    }
-                }
-                _ => panic!("Unsupported export type {value:?}"),
-            };
-
-            Ok((name, ext))
-            // tracing::debug!(?name, ?value, ?ty, "entry");
-        })
-        .collect()
 }
 
 #[derive(Debug, Clone)]
